changed the CustomerPersonal info Service(Calling the doApprove instead of save or update and validate the Customer documents against custDob)
git-svn-id: http://192.168.110.32/svn/Products/PFF/Base/Branches/Release-2.41.0@79151 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/main/java/com/pennant/backend/service/customermasters/impl/CustomerServiceImpl.java
@@ -51,7 +51,6 @@ import java.util.regex.Pattern;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.log4j.Logger;
-import org.springframework.beans.BeanUtils;
 
 import com.pennant.app.util.DateUtility;
 import com.pennant.app.util.ErrorUtil;
@@ -59,6 +58,7 @@ import com.pennant.app.util.SysParamUtil;
 import com.pennant.backend.dao.audit.AuditHeaderDAO;
 import com.pennant.backend.dao.collateral.ExtendedFieldRenderDAO;
 import com.pennant.backend.dao.customermasters.CustomerDAO;
+import com.pennant.backend.dao.customermasters.CustomerDocumentDAO;
 import com.pennant.backend.dao.customermasters.CustomerEmploymentDetailDAO;
 import com.pennant.backend.dao.systemmasters.IncomeTypeDAO;
 import com.pennant.backend.model.ErrorDetails;
@@ -66,6 +66,7 @@ import com.pennant.backend.model.audit.AuditDetail;
 import com.pennant.backend.model.audit.AuditHeader;
 import com.pennant.backend.model.customermasters.Customer;
 import com.pennant.backend.model.customermasters.CustomerDetails;
+import com.pennant.backend.model.customermasters.CustomerDocument;
 import com.pennant.backend.model.customermasters.CustomerEmploymentDetail;
 import com.pennant.backend.model.customermasters.WIFCustomer;
 import com.pennant.backend.service.GenericService;
@@ -89,6 +90,7 @@ public class CustomerServiceImpl extends GenericService<Customer> implements
 	private IncomeTypeDAO incomeTypeDAO;
 	private CustomerEmploymentDetailDAO customerEmploymentDetailDAO;
 	private ExtendedFieldRenderDAO extendedFieldRenderDAO;
+	private CustomerDocumentDAO customerDocumentDAO;
 
 
 	public CustomerServiceImpl() {
@@ -267,9 +269,9 @@ public class CustomerServiceImpl extends GenericService<Customer> implements
 			return auditHeader;
 		}
 
-		Customer customer = new Customer();
-		BeanUtils.copyProperties((Customer) auditHeader.getAuditDetail()
-				.getModelData(), customer);
+		Customer customer = (Customer) auditHeader.getAuditDetail().getModelData();
+		/*BeanUtils.copyProperties((Customer) auditHeader.getAuditDetail()
+				.getModelData(), customer);*/
 
 		if (customer.getRecordType().equals(PennantConstants.RECORD_TYPE_DEL)) {
 			tranType = PennantConstants.TRAN_DEL;
@@ -291,18 +293,23 @@ public class CustomerServiceImpl extends GenericService<Customer> implements
 				getCustomerDAO().update(customer, "");
 			}
 		}
+		if (auditHeader.getApiHeader() == null) {
+			getCustomerDAO().delete(customer, "_Temp");
+			auditHeader.setAuditTranType(PennantConstants.TRAN_WF);
+			getAuditHeaderDAO().addAudit(auditHeader);
+		}
 
-		getCustomerDAO().delete(customer, "_Temp");
-
-		auditHeader.setAuditTranType(PennantConstants.TRAN_WF);
+		/*auditHeader.setAuditTranType(PennantConstants.TRAN_WF);
 		String[] fields = PennantJavaUtil.getFieldDetails(new Customer(),"proceedToDedup,dedupFound,skipDedup");
 		auditHeader.setAuditDetail(new AuditDetail(auditHeader.getAuditTranType(), 1,fields[0],fields[1], customer.getBefImage(), customer));
-		getAuditHeaderDAO().addAudit(auditHeader);
+		getAuditHeaderDAO().addAudit(auditHeader);*/
 
 		auditHeader.setAuditTranType(tranType);
 		auditHeader.getAuditDetail().setAuditTranType(tranType);
-		auditHeader.setAuditDetail(new AuditDetail(auditHeader.getAuditTranType(), 1,fields[0],fields[1], customer.getBefImage(), customer));
+		auditHeader.getAuditDetail().setModelData(customer);
+
 		getAuditHeaderDAO().addAudit(auditHeader);
+
 		logger.debug("Leaving");
 		return auditHeader;
 	}
@@ -691,15 +698,33 @@ public class CustomerServiceImpl extends GenericService<Customer> implements
 		}
 		List<CustomerEmploymentDetail> customerEmploymentDetailList = customerEmploymentDetailDAO.
 										getCustomerEmploymentDetailsByID(customer.getCustID(), "");
-		for (CustomerEmploymentDetail custEmpDetails : customerEmploymentDetailList) {
-			if(custEmpDetails.getCustEmpFrom().before(customer.getCustDOB())){
-				String[] valueParm = new String[2];
-				valueParm[0] = "employment startDate:"+ DateUtility.formatDate(custEmpDetails.getCustEmpFrom(), 
-						PennantConstants.XMLDateFormat);
-				valueParm[1] = "Cust DOB:"+ DateUtility.formatDate(customer.getCustDOB(), 
-						PennantConstants.XMLDateFormat);
-				errorDetail = ErrorUtil.getErrorDetail(new ErrorDetails("30568", "", valueParm), "EN");
-				auditDetail.setErrorDetail(errorDetail);
+		if (customerEmploymentDetailList != null) {
+			for (CustomerEmploymentDetail custEmpDetails : customerEmploymentDetailList) {
+				if (custEmpDetails.getCustEmpFrom().before(customer.getCustDOB())) {
+					String[] valueParm = new String[2];
+					valueParm[0] = "employment startDate:"
+							+ DateUtility.formatDate(custEmpDetails.getCustEmpFrom(), PennantConstants.XMLDateFormat);
+					valueParm[1] = "Cust DOB:"
+							+ DateUtility.formatDate(customer.getCustDOB(), PennantConstants.XMLDateFormat);
+					errorDetail = ErrorUtil.getErrorDetail(new ErrorDetails("30568", "", valueParm), "EN");
+					auditDetail.setErrorDetail(errorDetail);
+				}
+			}
+		}
+		List<CustomerDocument> customerDocumentList = customerDocumentDAO.getCustomerDocumentByCustomerId(customer.getCustID());
+		if (customerDocumentList != null) {
+			for (CustomerDocument custDocDetails : customerDocumentList) {
+				if (custDocDetails.getCustDocIssuedOn() != null){
+				if (custDocDetails.getCustDocIssuedOn().before(customer.getCustDOB())) {
+					String[] valueParm = new String[2];
+					valueParm[0] = "CustDocIssuedOn:" + DateUtility.formatDate(custDocDetails.getCustDocIssuedOn(),
+							PennantConstants.XMLDateFormat);
+					valueParm[1] = "Cust DOB:"
+							+ DateUtility.formatDate(customer.getCustDOB(), PennantConstants.XMLDateFormat);
+					errorDetail = ErrorUtil.getErrorDetail(new ErrorDetails("30568", "", valueParm), "EN");
+					auditDetail.setErrorDetail(errorDetail);
+				}
+				}
 			}
 		}
 		logger.debug("Leaving");
@@ -766,5 +791,8 @@ public class CustomerServiceImpl extends GenericService<Customer> implements
 	}
 	public void setExtendedFieldRenderDAO(ExtendedFieldRenderDAO extendedFieldRenderDAO) {
 		this.extendedFieldRenderDAO = extendedFieldRenderDAO;
+	}
+	public void setCustomerDocumentDAO(CustomerDocumentDAO customerDocumentDAO) {
+		this.customerDocumentDAO = customerDocumentDAO;
 	}
 }
